Avoid XXE in clojure.data.xml
XXE (an *X*ML e*x*ternal *e*ntity attack) is possible in
clojure.data.xml versions 0.0.8 and below. We are unable to easily
upgrade to a more recent version (and no more recent "stable" version
exists).

We are introducing a wrapper namespace wrapping functions in
`clojure.data.xml` with safer XML parsing functions that apply XXE
protection by default, with the goal of ultimately removing the wrapper
namespace when the application is updated.